Refactor logger interface
