+++ pkg/devspace/cloud/util.go
@@ -107,7 +107,7 @@ func (p *Provider) PrintSpaces(cluster, name string, all bool) error {
 	}
 
 	if len(values) > 0 {
-		log.PrintTable(headerColumnNames, values)
+		log.PrintTable(log.GetInstance(), headerColumnNames, values)
 	} else {
 		log.Info("No spaces found")
 	}
